# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DIMO Device Data API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v2/vehicles/{tokenID}/analytics/daily-distance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get kilometers driven for a userDeviceID each day.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user device id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "IANAS time zone id, e.g., America/Los_Angeles",
                        "name": "time_zone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.DailyDistanceResp"
                        }
                    },
                    "404": {
                        "description": "no device found for user with provided parameters"
                    }
                }
            }
        },
        "/v2/vehicles/{tokenID}/analytics/total-distance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get kilometers driven for a userDeviceID since connected (ie. since we have data available)\nif it returns 0 for distanceDriven it means we have no odometer data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user device id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "no device found for user with provided parameters"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_controllers.DailyDistanceDay": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                }
            }
        },
        "internal_controllers.DailyDistanceResp": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_controllers.DailyDistanceDay"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}